// src/app/feature-modules/main/sidebarconfigure/groups/data/constant-value.enum.ts
export enum Constant {
    CONTROLLER = "/SymGroups",
    PATH = "/main/configure/groups",
    UNCLICKABLE = "unclickable",
    CLICKABLE = "clickable",
    CREATE = "/create",
    EDIT = "/edit",
    DELETE = "/delete",
    SORT_BY = "node_group_id",
    SORT_ORDER = "asc",
    FILTER_BY = "id",
    SAVE = "Save",
    UPDATE = "Update"
}